Move NewTestDate file.
